move module info into pkg/plugins
