+++ pkg/plugins/pluginassets/testdata/module-hash-no-manifest-txt/module.js
