+++ pkg/plugins/pluginassets/testdata/module-hash-no-manifest-txt/plugin.json
