+++ pkg/plugins/pluginassets/testdata/module-hash-no-module-js/plugin.json
