+++ pkg/plugins/pluginassets/testdata/module-hash-no-module-js/something.js
