+++ pkg/plugins/pluginassets/testdata/module-hash-valid-deeply-nested/MANIFEST.txt
