+++ pkg/plugins/pluginassets/testdata/module-hash-valid-deeply-nested/datasource/panels/one/plugin.json
