+++ pkg/plugins/pluginassets/testdata/module-hash-valid-deeply-nested/datasource/plugin.json
