+++ pkg/plugins/pluginassets/testdata/module-hash-valid-deeply-nested/module.js
