+++ pkg/plugins/pluginassets/testdata/module-hash-valid-deeply-nested/plugin.json
