+++ pkg/plugins/pluginassets/testdata/module-hash-valid-nested/MANIFEST.txt
