+++ pkg/plugins/pluginassets/testdata/module-hash-valid-nested/datasource/module.js
