+++ pkg/plugins/pluginassets/testdata/module-hash-valid-nested/datasource/plugin.json
