+++ pkg/plugins/pluginassets/testdata/module-hash-valid-nested/module.js
