+++ pkg/plugins/pluginassets/testdata/module-hash-valid-nested/panels/one/plugin.json
