+++ pkg/plugins/pluginassets/testdata/module-hash-valid-nested/plugin.json
