+++ pkg/plugins/pluginassets/testdata/module-hash-valid/MANIFEST.txt
