+++ pkg/plugins/pluginassets/testdata/module-hash-valid/module.js
