+++ pkg/plugins/pluginassets/testdata/module-hash-valid/plugin.json
