+++ pkg/services/pluginsintegration/pluginsintegration.go
@@ -131,7 +131,9 @@ var WireSet = wire.NewSet(
 	plugincontext.ProvideBaseService,
 	wire.Bind(new(plugincontext.BasePluginContextProvider), new(*plugincontext.BaseProvider)),
 	plugininstaller.ProvideService,
-	pluginassets.ProvideService,
+	pluginassets2.ProvideService,
+	wire.Bind(new(plugins.PluginModuleInfo), new(*pluginassets2.Service)),
+	pluginassets.NewService,
 	pluginchecker.ProvidePreinstall,
 	wire.Bind(new(pluginchecker.Preinstall), new(*pluginchecker.PreinstallImpl)),
 	advisor.ProvideService,
